Start clean up of visual style model - remove code that is never called - export all visual style models to the top level of the visual style module - make the visual style fn impl export consistent with other model exports
