+++ src/models/VisualStyleModel/VisualMappingFunction/index.ts
@@ -1,5 +1,6 @@
-export { DiscreteMappingFunction } from './DiscreteMappingFunction'
-export { ContinuousMappingFunction } from './ContinuousMappingFunction'
-export { PassthroughMappingFunction } from './PassthroughMappingFunction'
-export { MappingFunctionType } from './MappingFunctionType'
-export { VisualMappingFunction } from './VisualMappingFunction'
+export * from './DiscreteMappingFunction'
+export * from './ContinuousMappingFunction'
+export * from './PassthroughMappingFunction'
+export * from './MappingFunctionType'
+export * from './VisualMappingFunction'
+export * from './Mapper'
