+++ src/models/VisualStyleModel/impl/VisualStyleFnImpl.ts
@@ -1,17 +1,38 @@
-import { Aspect } from '../../../utils/cx/Cx2/Aspect'
-import { CxDescriptor } from '../../../utils/cx/Cx2/CxDescriptor'
-import { MetaData } from '../../../utils/cx/Cx2/MetaData'
-import { Status } from '../../../utils/cx/Cx2/Status'
+import _ from 'lodash'
+import { Cx2 } from '../../../utils/cx/Cx2'
+import * as cxUtil from '../../../utils/cx/cx2-util'
+
 import { NetworkView } from '../../ViewModel'
-import { VisualProperty } from '../VisualProperty'
-import { VisualPropertyGroup } from '../VisualPropertyGroup'
-import { VisualPropertyValueType } from '../VisualPropertyValue'
-import { VisualStyle } from '../VisualStyle'
-import { NetworkViewSources, VisualStyleFn } from '../VisualStyleFn'
+
+import { ValueType } from '../../TableModel'
+
+import {
+  VisualStyle,
+  VisualPropertyName,
+  VisualPropertyGroup,
+  ContinuousFunctionControlPoint,
+  VisualPropertyValueType,
+  VisualProperty,
+  Bypass,
+  NetworkViewSources,
+  DiscreteMappingFunction,
+  ContinuousMappingFunction,
+  PassthroughMappingFunction,
+  MappingFunctionType,
+} from '..'
+
+import {
+  CXId,
+  CXVisualMappingFunction,
+  cxVisualPropertyConverter,
+  CXVisualPropertyConverter,
+  CXVisualPropertyValue,
+} from './cxVisualPropertyConverter'
+
 import { getDefaultVisualStyle } from './DefaultVisualStyle'
 import { createNewNetworkView, updateNetworkView } from './visualStyle-utils'
 
-const applyVisualStyle = (data: NetworkViewSources): NetworkView => {
+export const applyVisualStyle = (data: NetworkViewSources): NetworkView => {
   const { network, visualStyle, nodeTable, edgeTable, networkView } = data
 
   if (networkView !== undefined) {
@@ -27,7 +48,7 @@ const applyVisualStyle = (data: NetworkViewSources): NetworkView => {
   }
 }
 
-const nodeVisualProperties = (
+export const nodeVisualProperties = (
   visualStyle: VisualStyle,
 ): Array<VisualProperty<VisualPropertyValueType>> => {
   return Object.values(visualStyle).filter(
@@ -35,7 +56,7 @@ const nodeVisualProperties = (
   )
 }
 
-const edgeVisualProperties = (
+export const edgeVisualProperties = (
   visualStyle: VisualStyle,
 ): Array<VisualProperty<VisualPropertyValueType>> => {
   return Object.values(visualStyle).filter(
@@ -43,7 +64,7 @@ const edgeVisualProperties = (
   )
 }
 
-const networkVisualProperties = (
+export const networkVisualProperties = (
   visualStyle: VisualStyle,
 ): Array<VisualProperty<VisualPropertyValueType>> => {
   return Object.values(visualStyle).filter(
@@ -51,24 +72,275 @@ const networkVisualProperties = (
   )
 }
 
-const createVisualStyle = (): VisualStyle => {
+export const createVisualStyle = (): VisualStyle => {
   // create new copy of the default style instead of returning the same instance
   return getDefaultVisualStyle()
 }
 
-export const VisualStyleFnImpl: VisualStyleFn = {
-  createVisualStyle,
-  createVisualStyleFromCx: function (
-    cx:
-      | [CxDescriptor, ...Aspect[], Status]
-      | [CxDescriptor, ...Aspect[], MetaData, Status]
-      | [CxDescriptor, MetaData, ...Aspect[], Status]
-      | [CxDescriptor, MetaData, ...Aspect[], MetaData, Status],
-  ): VisualStyle {
-    throw new Error('Function not implemented.')
-  },
-  applyVisualStyle,
-  nodeVisualProperties,
-  edgeVisualProperties,
-  networkVisualProperties,
+// convert cx visual properties to app visual style model
+export const createVisualStyleFromCx = (cx: Cx2): VisualStyle => {
+  const visualStyle: VisualStyle = createVisualStyle()
+  const visualProperties = cxUtil.getVisualProperties(cx)
+  const nodeBypasses = cxUtil.getNodeBypasses(cx) ?? []
+  const edgeBypasses = cxUtil.getEdgeBypasses(cx) ?? []
+  const defaultNodeProperties =
+    visualProperties.visualProperties[0]?.default?.node ?? {}
+  const defaultEdgeProperties =
+    visualProperties.visualProperties[0]?.default?.edge ?? {}
+  const defaultNetworkProperties =
+    visualProperties.visualProperties[0]?.default?.network ?? {}
+  const nodeMapping = visualProperties.visualProperties[0]?.nodeMapping ?? {}
+  const edgeMapping = visualProperties.visualProperties[0]?.edgeMapping ?? {}
+
+  const nodeBypassMap: Map<
+    VisualPropertyName,
+    Bypass<VisualPropertyValueType>
+  > = new Map()
+  const edgeBypassMap: Map<
+    VisualPropertyName,
+    Bypass<VisualPropertyValueType>
+  > = new Map()
+
+  // group bypasses by visual property instead of by element
+  nodeBypasses?.nodeBypasses?.forEach(
+    (entry: { id: CXId; v: Record<string, object> }) => {
+      const { id, v } = entry
+      Object.keys(v).forEach((cxVPName) => {
+        const entry = Object.entries(cxVisualPropertyConverter).find(
+          ([vpName, cxVPConverter]) => cxVPConverter.cxVPName === cxVPName,
+        )
+
+        if (entry != null) {
+          const [vpName, cxVPConverter] = entry as [
+            VisualPropertyName,
+            CXVisualPropertyConverter<VisualPropertyValueType>,
+          ]
+
+          if (nodeBypassMap.has(vpName)) {
+            const entry = nodeBypassMap.get(vpName) ?? new Map()
+            entry.set(
+              String(id),
+              cxVPConverter.valueConverter(
+                v[cxVPName] as CXVisualPropertyValue,
+              ),
+            )
+            nodeBypassMap.set(vpName, entry)
+          } else {
+            nodeBypassMap.set(vpName, new Map())
+          }
+        }
+      })
+    },
+  )
+
+  // group bypasses by visual property instead of by element
+  edgeBypasses?.nodeBypasses?.forEach(
+    (entry: { id: CXId; v: Record<string, object> }) => {
+      const { id, v } = entry
+      Object.keys(v).forEach((cxVPName) => {
+        const entry = Object.entries(cxVisualPropertyConverter).find(
+          ([vpName, cxVPConverter]) => cxVPConverter.cxVPName === cxVPName,
+        )
+
+        if (entry != null) {
+          const [vpName, cxVPConverter] = entry as [
+            VisualPropertyName,
+            CXVisualPropertyConverter<VisualPropertyValueType>,
+          ]
+
+          if (edgeBypassMap.has(vpName)) {
+            const entry = edgeBypassMap.get(vpName) ?? new Map()
+            entry.set(
+              String(id),
+              cxVPConverter.valueConverter(
+                v[cxVPName] as CXVisualPropertyValue,
+              ),
+            )
+            edgeBypassMap.set(vpName, entry)
+          } else {
+            edgeBypassMap.set(vpName, new Map())
+          }
+        }
+      })
+    },
+  )
+
+  const vpGroups = [
+    {
+      vps: nodeVisualProperties(visualStyle),
+      getDefault: (cxVPName: string) => defaultNodeProperties[cxVPName],
+      getMapping: (
+        cxVPName: string,
+      ): CXVisualMappingFunction<CXVisualPropertyValue> | null =>
+        nodeMapping[cxVPName] as CXVisualMappingFunction<CXVisualPropertyValue>,
+      getBypass: (): Map<VisualPropertyName, Bypass<VisualPropertyValueType>> =>
+        nodeBypassMap,
+    },
+    {
+      vps: edgeVisualProperties(visualStyle),
+      getDefault: (cxVPName: string) => defaultEdgeProperties[cxVPName],
+      getMapping: (
+        cxVPName: string,
+      ): CXVisualMappingFunction<CXVisualPropertyValue> | null =>
+        edgeMapping?.[
+          cxVPName
+        ] as CXVisualMappingFunction<CXVisualPropertyValue>,
+      getBypass: (): Map<VisualPropertyName, Bypass<VisualPropertyValueType>> =>
+        edgeBypassMap,
+    },
+    {
+      vps: networkVisualProperties(visualStyle),
+      getDefault: (cxVPName: string) => defaultNetworkProperties[cxVPName],
+      getMapping: (cxVPName: string) => null,
+      getBypass: () => new Map(), // no mappings or bypasses for network vps
+    },
+  ]
+
+  vpGroups.forEach((group) => {
+    const { vps, getDefault, getMapping, getBypass } = group
+    vps.forEach((vp: VisualProperty<VisualPropertyValueType>) => {
+      const { name: vpName } = vp
+      const converter = cxVisualPropertyConverter[vpName]
+
+      const isSupportedCXProperty = converter != null
+
+      if (isSupportedCXProperty) {
+        const cxDefault = getDefault(
+          converter.cxVPName,
+        ) as CXVisualPropertyValue
+        const cxMapping = getMapping(converter.cxVPName)
+        const cxBypass = getBypass()
+
+        if (cxDefault != null) {
+          visualStyle[vpName].defaultValue = converter.valueConverter(cxDefault)
+        }
+
+        if (cxMapping != null) {
+          switch (cxMapping.type) {
+            case 'PASSTHROUGH': {
+              const m: PassthroughMappingFunction = {
+                type: MappingFunctionType.Passthrough,
+                visualPropertyType: vp.type,
+                attribute: cxMapping.definition.attribute,
+                defaultValue: vp.defaultValue,
+              }
+              visualStyle[vpName].mapping = m
+              break
+            }
+            case 'DISCRETE': {
+              const vpValueMap = new Map()
+              cxMapping.definition.map.forEach((mapEntry) => {
+                const { v, vp } = mapEntry
+                vpValueMap.set(v, converter.valueConverter(vp))
+              })
+              const m: DiscreteMappingFunction = {
+                type: MappingFunctionType.Discrete,
+                attribute: cxMapping.definition.attribute,
+                vpValueMap,
+                visualPropertyType: vp.type,
+                defaultValue: vp.defaultValue,
+              }
+              visualStyle[vpName].mapping = m
+              break
+            }
+            case 'CONTINUOUS': {
+              const numMapEntries = cxMapping.definition.map.length
+              if (numMapEntries < 2) {
+                visualStyle[vpName].mapping = undefined
+                break
+              }
+
+              let min = null
+              let max = null
+
+              if (
+                cxMapping.definition.map[0].max != null &&
+                cxMapping.definition.map[0].maxVPValue != null
+              ) {
+                min = {
+                  value: cxMapping.definition.map[0].max as ValueType,
+                  vpValue: converter.valueConverter(
+                    cxMapping.definition.map[0].maxVPValue,
+                  ),
+                  inclusive: cxMapping.definition.map[0].includeMax,
+                }
+              }
+
+              if (
+                cxMapping.definition.map[numMapEntries - 1].min != null &&
+                cxMapping.definition.map[numMapEntries - 1].minVPValue != null
+              ) {
+                max = {
+                  value: cxMapping.definition.map[numMapEntries - 1]
+                    .min as ValueType,
+                  vpValue: converter.valueConverter(
+                    cxMapping.definition.map[numMapEntries - 1]
+                      .minVPValue as CXVisualPropertyValue,
+                  ),
+                  inclusive:
+                    cxMapping.definition.map[numMapEntries - 1].includeMin,
+                }
+              }
+
+              const controlPoints: ContinuousFunctionControlPoint[] = []
+
+              // only iterate through the middle entries of the map
+              // i.e. exclue min and max
+              for (let i = 1; i <= numMapEntries - 2; i++) {
+                const mapEntry = cxMapping.definition.map[i]
+                if (mapEntry.minVPValue != null && mapEntry.min != null) {
+                  controlPoints.push({
+                    value: mapEntry.min as ValueType,
+                    vpValue: converter.valueConverter(mapEntry.minVPValue),
+                  })
+                }
+
+                if (mapEntry.maxVPValue != null && mapEntry.max != null) {
+                  controlPoints.push({
+                    value: mapEntry.max as ValueType,
+                    vpValue: converter.valueConverter(mapEntry.maxVPValue),
+                  })
+                }
+              }
+
+              const uniqueCtrlPts = _.uniqWith(controlPoints, _.isEqual)
+
+              const sortedCtrlPts = Array.from(uniqueCtrlPts).sort(
+                (a, b) => (a.value as number) - (b.value as number),
+              )
+
+              if (min != null && max != null && controlPoints.length > 0) {
+                const m: ContinuousMappingFunction = {
+                  type: MappingFunctionType.Continuous,
+                  attribute: cxMapping.definition.attribute,
+                  min,
+                  max,
+                  controlPoints: sortedCtrlPts,
+                  visualPropertyType: vp.type,
+                  defaultValue: vp.defaultValue,
+                }
+                visualStyle[vpName].mapping = m
+              } else {
+                // visualStyle[vpName].mapping = undefined
+              }
+              break
+            }
+            default:
+              break
+          }
+        }
+
+        visualStyle[vpName].bypassMap = cxBypass.get(vpName) ?? new Map()
+      } else {
+        // property is not found in cx, in theory all cytoscape web properties should be in
+        // cx, if this happens, it is a bug
+        console.error(`Property ${vpName} not found in CX`)
+      }
+    })
+
+    // some cx properties are probably not handled in this conversion,
+    // we should find a way to store them and then restore them when we round trip
+  })
+
+  return visualStyle
 }
